move magenta_rt/nnx/configs.py to magenta_rt/nnx/model.py
